fix: product image upload + discount decimal support
- Fix Vue reactive proxy breaking instanceof File check (rawImageFile)
- Add try-catch + error feedback for image processing failures
- Add webp,svg to accepted image mimes
- Change discount_percentage: integer → numeric/decimal(5,2)
- Add migration for DB column change

+++ app/Http/Resources/ProductResource.php
@@ -18,7 +18,7 @@ class ProductResource extends JsonResource
             'name' => $this->name,
             'description' => $this->description,
             'price'               => (float) $this->price,
-            'discount_percentage' => (int) $this->discount_percentage,
+            'discount_percentage' => (float) $this->discount_percentage,
             'discounted_price'    => $this->discount_percentage > 0
                 ? round($this->price * (1 - $this->discount_percentage / 100), 2)
                 : null,
